requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
fastAPI
snowflake-connector-python
typing-extensions
snowflake-sqlalchemy
python-multipart
python-jose[cryptography]
passlib[bcrypt]
jose